Reorganize files & folders to follow rubygems best practices regarding folder structure.
