+++ lib/mongoid/tracking/aggregates.rb
