+++ lib/mongoid/tracking/core_ext.rb
@@ -1,0 +1,3 @@
+# encoding: utf-8
+require 'mongoid/tracking/core_ext/time'
+require 'mongoid/tracking/core_ext/range'
