+++ lib/mongoid/tracking/core_ext/range.rb
