+++ lib/mongoid/tracking/tracker_aggregates.rb
