--- lib/trackoid/core_ext.rb
@@ -1,3 +1,0 @@
-# encoding: utf-8
-require 'trackoid/core_ext/time'
-require 'trackoid/core_ext/range'
